+++ src/zvode.f
@@ -1,4 +1,4 @@
-C Original authors: Peter N. Brown, Alan C. Hindmarsh, 
+C Original authors: Peter N. Brown, Alan C. Hindmarsh,
 C   Geore D. Byrne (see original author statement below)
 C
 C  Adapted for use in R package deSolve by the deSolve authors.
@@ -9,7 +9,7 @@ C
      1            ISTATE, IOPT, ZWORK, LZW, RWORK, LRW, IWORK, LIW,
      2            JAC, MF, RPAR, IPAR)
       EXTERNAL F, JAC
-      COMPLEX(KIND=8) Y, ZWORK
+      COMPLEX(KIND=KIND(0.0d0)) Y, ZWORK
       DOUBLE PRECISION T, TOUT, RTOL, ATOL, RWORK
       INTEGER NEQ, ITOL, ITASK, ISTATE, IOPT, LZW, LRW, IWORK, LIW,
      1        MF, IPAR
@@ -1744,7 +1744,7 @@ C----------------------- End of Subroutine ZVODE -----------------------
       SUBROUTINE ZVHIN (N, T0, Y0, YDOT, F, RPAR, IPAR, TOUT, UROUND,
      1   EWT, ITOL, ATOL, Y, TEMP, H0, NITER, IER)
       EXTERNAL F
-      COMPLEX(KIND=8) Y0, YDOT, Y, TEMP
+      COMPLEX(KIND=KIND(0.0d0)) Y0, YDOT, Y, TEMP
       DOUBLE PRECISION T0, TOUT, UROUND, EWT, ATOL, H0
       INTEGER N, IPAR, ITOL, NITER, IER
       DIMENSION Y0(*), YDOT(*), EWT(*), ATOL(*), Y(*),
@@ -1841,7 +1841,7 @@ C Estimate the second derivative as a difference quotient in f. --------
       CALL F (N, T1, Y, TEMP, RPAR, IPAR)
       DO 70 I = 1, N
         TEMP(I) = (TEMP(I) - YDOT(I))/H
- 70   CONTINUE 
+ 70   CONTINUE
       YDDNRM = ZVNORM (N, TEMP, EWT)
 C Get the corresponding new value of h. --------------------------------
       IF (YDDNRM*HUB*HUB .GT. TWO) THEN
@@ -1882,7 +1882,7 @@ C----------------------- End of Subroutine ZVHIN -----------------------
       END
 *DECK ZVINDY
       SUBROUTINE ZVINDY (T, K, YH, LDYH, DKY, IFLAG)
-      COMPLEX(KIND=8) YH, DKY
+      COMPLEX(KIND=KIND(0.0d0)) YH, DKY
       DOUBLE PRECISION T
       INTEGER K, LDYH, IFLAG
       DIMENSION YH(LDYH,*), DKY(*)
@@ -1985,7 +1985,7 @@ C
         JJ1 = JP1 - K
         DO 30 JJ = JJ1, J
           IC = IC*JJ
- 30     CONTINUE 
+ 30     CONTINUE
  35     C = REAL(IC)
         DO 40 I = 1, N
           DKY(I) = C*YH(I,JP1) + S*DKY(I)
@@ -2012,7 +2012,7 @@ C----------------------- End of Subroutine ZVINDY ----------------------
       SUBROUTINE ZVSTEP (Y, YH, LDYH, YH1, EWT, SAVF, VSAV, ACOR,
      1                  WM, IWM, F, JAC, PSOL, VNLS, RPAR, IPAR)
       EXTERNAL F, JAC, PSOL, VNLS
-      COMPLEX(KIND=8) Y, YH, YH1, SAVF, VSAV, ACOR, WM
+      COMPLEX(KIND=KIND(0.0d0)) Y, YH, YH1, SAVF, VSAV, ACOR, WM
       DOUBLE PRECISION EWT
       INTEGER LDYH, IWM, IPAR
       DIMENSION Y(*), YH(LDYH,*), YH1(*), EWT(*), SAVF(*), VSAV(*),
@@ -2667,7 +2667,7 @@ C----------------------- End of Subroutine ZVSET -----------------------
       END
 *DECK ZVJUST
       SUBROUTINE ZVJUST (YH, LDYH, IORD)
-      COMPLEX(KIND=8) YH
+      COMPLEX(KIND=KIND(0.0d0)) YH
       INTEGER LDYH, IORD
       DIMENSION YH(LDYH,*)
 C-----------------------------------------------------------------------
@@ -2845,7 +2845,7 @@ C----------------------- End of Subroutine ZVJUST ----------------------
       SUBROUTINE ZVNLSD (Y, YH, LDYH, VSAV, SAVF, EWT, ACOR, IWM, WM,
      1                 F, JAC, PDUM, NFLAG, RPAR, IPAR)
       EXTERNAL F, JAC, PDUM
-      COMPLEX(KIND=8) Y, YH, VSAV, SAVF, ACOR, WM
+      COMPLEX(KIND=KIND(0.0d0)) Y, YH, VSAV, SAVF, ACOR, WM
       DOUBLE PRECISION EWT
       INTEGER LDYH, IWM, NFLAG, IPAR
       DIMENSION Y(*), YH(LDYH,*), VSAV(*), SAVF(*), EWT(*), ACOR(*),
@@ -3081,7 +3081,7 @@ C----------------------- End of Subroutine ZVNLSD ----------------------
       SUBROUTINE ZVJAC (Y, YH, LDYH, EWT, FTEM, SAVF, WM, IWM, F, JAC,
      1                 IERPJ, RPAR, IPAR)
       EXTERNAL F, JAC
-      COMPLEX(KIND=8) Y, YH, FTEM, SAVF, WM
+      COMPLEX(KIND=KIND(0.0d0)) Y, YH, FTEM, SAVF, WM
       DOUBLE PRECISION EWT
       INTEGER LDYH, IWM, IERPJ, IPAR
       DIMENSION Y(*), YH(LDYH,*), EWT(*), FTEM(*), SAVF(*),
@@ -3157,7 +3157,7 @@ C
 C
 C Type declarations for local variables --------------------------------
 C
-      COMPLEX(KIND=8) DI, R1, YI, YJ, YJJ
+      COMPLEX(KIND=KIND(0.0d0)) DI, R1, YI, YJ, YJJ
       DOUBLE PRECISION CON, FAC, ONE, PT1, R, R0, THOU, ZERO
       INTEGER I, I1, I2, IER, II, J, J1, JJ, JOK, LENP, MBA, MBAND,
      1        MEB1, MEBAND, ML, ML1, MU, NP1
@@ -3316,7 +3316,7 @@ C If MITER = 5, make ML+MU+1 calls to F to approximate the Jacobian. ---
           YI = Y(I)
           R = MAX(SRUR*ABS(YI),R0/EWT(I))
           Y(I) = Y(I) + R
- 530    CONTINUE 
+ 530    CONTINUE
         CALL F (N, TN, Y, FTEM, RPAR, IPAR)
         DO 550 JJ = J,N,MBAND
           Y(JJ) = YH(JJ,1)
@@ -3359,7 +3359,7 @@ C----------------------- End of Subroutine ZVJAC -----------------------
       END
 *DECK ZACOPY
       SUBROUTINE ZACOPY (NROW, NCOL, A, NROWA, B, NROWB)
-      COMPLEX(KIND=8) A, B
+      COMPLEX(KIND=KIND(0.0d0)) A, B
       INTEGER NROW, NCOL, NROWA, NROWB
       DIMENSION A(NROWA,NCOL), B(NROWB,NCOL)
 C-----------------------------------------------------------------------
@@ -3385,7 +3385,7 @@ C----------------------- End of Subroutine ZACOPY ----------------------
       END
 *DECK ZVSOL
       SUBROUTINE ZVSOL (WM, IWM, X, IERSL)
-      COMPLEX(KIND=8) WM, X
+      COMPLEX(KIND=KIND(0.0d0)) WM, X
       INTEGER IWM, IERSL
       DIMENSION WM(*), IWM(*), X(*)
 C-----------------------------------------------------------------------
@@ -3428,7 +3428,7 @@ C
 C
 C Type declarations for local variables --------------------------------
 C
-      COMPLEX(KIND=8) DI
+      COMPLEX(KIND=KIND(0.0d0)) DI
       DOUBLE PRECISION ONE, PHRL1, R, ZERO
       INTEGER I, MEBAND, ML, MU
 C-----------------------------------------------------------------------
@@ -3574,7 +3574,7 @@ C***ROUTINES CALLED  (NONE)
 C***REVISION HISTORY  (YYMMDD)
 C   060502  DATE WRITTEN, modified from DEWSET of 930809.
 C***END PROLOGUE  ZEWSET
-      COMPLEX(KIND=8) YCUR
+      COMPLEX(KIND=KIND(0.0d0)) YCUR
       DOUBLE PRECISION RTOL, ATOL, EWT
       INTEGER N, ITOL
       INTEGER I
@@ -3633,7 +3633,7 @@ C***ROUTINES CALLED  ZABSSQ
 C***REVISION HISTORY  (YYMMDD)
 C   060502  DATE WRITTEN, modified from DVNORM of 930809.
 C***END PROLOGUE  ZVNORM
-      COMPLEX(KIND=8) V
+      COMPLEX(KIND=KIND(0.0d0)) V
       DOUBLE PRECISION W,   SUM, ZABSSQ
       INTEGER N,   I
       DIMENSION V(N), W(N)
@@ -3662,8 +3662,8 @@ C    ZABSSQ = DREAL(Z)**2 * DIMAG(Z)**2
 C***REVISION HISTORY  (YYMMDD)
 C   060502  DATE WRITTEN.
 C***END PROLOGUE  ZABSSQ
-      COMPLEX(KIND=8) Z
-      ZABSSQ = DREAL(Z)**2 + DIMAG(Z)**2
+      COMPLEX(KIND=KIND(0.0d0)) Z
+      ZABSSQ = REAL(Z)**2 + AIMAG(Z)**2
       RETURN
 C----------------------- END OF FUNCTION ZABSSQ ------------------------
       END
@@ -3680,7 +3680,7 @@ C  Minor modification of BLAS routine ZSCAL.
 C***REVISION HISTORY  (YYMMDD)
 C   060530  DATE WRITTEN.
 C***END PROLOGUE  DZSCAL
-      COMPLEX(KIND=8) ZX(*)
+      COMPLEX(KIND=KIND(0.0d0)) ZX(*)
       DOUBLE PRECISION DA
       INTEGER I,INCX,IX,N
 C
@@ -3711,7 +3711,7 @@ C  Minor modification of BLAS routine ZAXPY.
 C***REVISION HISTORY  (YYMMDD)
 C   060530  DATE WRITTEN.
 C***END PROLOGUE  DZAXPY
-      COMPLEX(KIND=8) ZX(*),ZY(*)
+      COMPLEX(KIND=KIND(0.0d0)) ZX(*),ZY(*)
       DOUBLE PRECISION DA
       INTEGER I,INCX,INCY,IX,IY,N
       IF(N.LE.0)RETURN
@@ -3738,7 +3738,7 @@ C Code for both increments equal to 1
 
       subroutine zgesl(a,lda,n,ipvt,b,job)
       integer lda,n,ipvt(1),job
-      COMPLEX(KIND=8) a(lda,*),b(*)
+      COMPLEX(KIND=KIND(0.0d0)) a(lda,*),b(*)
 c
 c     zgesl solves the COMPLEX(KIND=8) system
 c     a * x = b  or  ctrans(a) * x = b
@@ -3797,11 +3797,11 @@ c     fortran dconjg
 c
 c     internal variables
 c
-      COMPLEX(KIND=8) zdotc,t
+      COMPLEX(KIND=KIND(0.0d0)) zdotc,t
       integer k,kb,l,nm1
 C KS      double precision dreal,dimag
 C KS      COMPLEX(KIND=8) zdumr,zdumi
-C KS     dreal(zdumr) = zdumr
+C KS      dreal(zdumr) = zdumr
 C KS      dimag(zdumi) = (0.0d0,-1.0d0)*zdumi
 c
       nm1 = n - 1
@@ -3838,7 +3838,7 @@ c        first solve  ctrans(u)*y = b
 c
          do 60 k = 1, n
             t = zdotc(k-1,a(1,k),1,b(1),1)
-            b(k) = (b(k) - t)/dconjg(a(k,k))
+            b(k) = (b(k) - t)/conjg(a(k,k))
    60    continue
 c
 c        now solve ctrans(l)*x = y
@@ -3861,7 +3861,7 @@ c
 
       subroutine zgbfa(abd,lda,n,ml,mu,ipvt,info)
       integer lda,n,ml,mu,ipvt(*),info
-      COMPLEX(KIND=8) abd(lda,*)
+      COMPLEX(KIND=KIND(0.0d0)) abd(lda,*)
 c
 c     zgbfa factors a COMPLEX(KIND=8) band matrix by elimination.
 c
@@ -3945,7 +3945,7 @@ c     fortran dabs,max0,min0
 c
 c     internal variables
 c
-      COMPLEX(KIND=8) t
+      COMPLEX(KIND=KIND(0.0d0)) t
       integer i,izamax,i0,j,ju,jz,j0,j1,k,kp1,l,lm,m,mm,nm1
 c
 CKS      COMPLEX(KIND=8) zdum
@@ -4043,7 +4043,7 @@ c
 
       subroutine zgbsl(abd,lda,n,ml,mu,ipvt,b,job)
       integer lda,n,ml,mu,ipvt(1),job
-      COMPLEX(KIND=8) abd(lda,*),b(*)
+      COMPLEX(KIND=KIND(0.0d0)) abd(lda,*),b(*)
 c
 c     zgbsl solves the COMPLEX(KIND=8) band system
 c     a * x = b  or  ctrans(a) * x = b
@@ -4108,7 +4108,7 @@ c     fortran dconjg,min0
 c
 c     internal variables
 c
-      COMPLEX(KIND=8) zdotc,t
+      COMPLEX(KIND=KIND(0.0d0)) zdotc,t
       integer k,kb,l,la,lb,lm,m,nm1
 C      double precision dreal,dimag
 C      COMPLEX(KIND=8) zdumr,zdumi
@@ -4158,7 +4158,7 @@ c
             la = m - lm
             lb = k - lm
             t = zdotc(lm,abd(la,k),1,b(lb),1)
-            b(k) = (b(k) - t)/dconjg(abd(m,k))
+            b(k) = (b(k) - t)/conjg(abd(m,k))
    60    continue
 c
 c        now solve ctrans(l)*x = y
@@ -4182,20 +4182,17 @@ c
       end
 C KARLINE: created true functions out of these statement functions
 C Thomas:  removed function definitions for dreal and dimag,
-C          they were already existing.
-C     We may consider to use 'real' and 'imag' consistently for
-C        future versions.
-C Thomas 2023: recent standard is aimag
+C          as 'real' and 'aimag' are now standards.
       double precision function cabs1(zdum)
-      complex (kind = 8), intent (in) :: zdum
+      COMPLEX(KIND=KIND(0.0d0)) zdum
         cabs1 = dabs(real(zdum)) + dabs(aimag(zdum))
-      end function 
+      end function
 C KARLINE: end new functions
 
 
       subroutine zgefa(a,lda,n,ipvt,info)
       integer lda,n,ipvt(*),info
-      COMPLEX(KIND=8) a(lda,*)
+      COMPLEX(KIND=KIND(0.0d0)) a(lda,*)
 c
 c     zgefa factors a COMPLEX(KIND=8) matrix by gaussian elimination.
 c
@@ -4243,13 +4240,13 @@ c     fortran dabs
 c
 c     internal variables
 c
-      COMPLEX(KIND=8) t
+      COMPLEX(KIND=KIND(0.0d0)) t
       integer izamax,j,k,kp1,l,nm1
 c
-C KS      COMPLEX(KIND=8) zdum
+C KS    COMPLEX(KIND=8) zdum
       double precision cabs1
-C      double precision dreal,dimag
-C KS      COMPLEX(KIND=8) zdumr,zdumi
+C     double precision dreal,dimag
+C KS    COMPLEX(KIND=8) zdumr,zdumi
 C Karline: next three statement functions replaced with true functions above
 C      dreal(zdumr) = zdumr
 C      dimag(zdumi) = (0.0d0,-1.0d0)*zdumi